--- a/examples/snake.nk
+++ b/examples/snake.nk
@@ -47,8 +47,8 @@ of nested block calls. Phew!
   ] @: opp?
 
   [ "( -- ): set X, Y to random values in screen bounds."
-    0 console:width  randFromTo =: x
-    0 console:height randFromTo =: y
+    0 console:width  1 - randFromTo =: x
+    0 console:height 1 - randFromTo =: y
   ] @: random
 
   [